Define null-ls as a module / split logic across files
